refactor: organizate folders
