+++ packages/server/src/controllers/HealthController/index.ts
@@ -1,7 +1,7 @@
-import {Request, Response} from 'express'
+import { Request, Response } from 'express'
 
 const HealthController = (request: Request, response: Response) => {
-    return response.status(200).json({message: 'ok'})
+    return response.status(200).json({ error: false, message: 'Server is fine 🔥' })
 }
 
 export default HealthController
