+++ packages/server/src/controllers/UploadController/index.ts
@@ -1,0 +1,5 @@
+import TransactionFile from "./TransactionFile"
+
+export default {
+    TransactionFile
+}
