+++ packages/server/src/routes/upload/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import UploadTransactionFileController from '../../controllers/UploadTransactionFileController'
+import UploadController from '../../controllers/UploadController'
 import { storage } from "../../middleware/upload";
 import multer from "multer";
 
@@ -9,6 +9,6 @@ const upload = multer({
 
 const uploadRoutes = Router()
 
-uploadRoutes.post('/transaction-file', upload.single(''), UploadTransactionFileController)
+uploadRoutes.post('/transaction-file', upload.single(''), UploadController.TransactionFile)
 
 export { uploadRoutes }
